Experiments

+++ source/aogmaneo/encoder.h
@@ -69,7 +69,7 @@ private:
         const Int2 &column_pos,
         const Int_Buffer* input_cis,
         int vli,
-        unsigned int* state,
+        unsigned long* state,
         const Params &params
     );
